requirements.txt
adamp==0.3.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autograd==1.8.0
autograd-gamma==0.5.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi==2025.4.26
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.3.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
efficientnet_pytorch==0.7.1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock @ file:///croot/filelock_1744281381737/work
fonttools==4.58.0
formulaic==1.1.1
fqdn==1.5.1
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///croot/gmpy2_1738085463648/work
h11==0.16.0
h5py==3.11.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.3
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
interface-meta==1.3.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
lifelines==0.30.0
lightning-utilities==0.15.2
loralib==0.1.2
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.1.3
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1737039604450/work
notebook==7.4.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-nccl-cu12==2.26.5
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
overrides==7.7.0
packaging==25.0
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.8
prometheus_client==0.22.0
prompt_toolkit==3.0.51
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.4.0
referencing==0.36.2
requests @ file:///croot/requests_1730999120400/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
rpy2 @ file:///home/conda/feedstock_root/build_artifacts/rpy2_1696426032385/work
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.14.0
Send2Trash==1.8.3
sentry-sdk==2.41.0
simplegeneric @ file:///home/conda/feedstock_root/build_artifacts/simplegeneric_1734685420077/work
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.4
sympy @ file:///croot/sympy_1738108488918/work
tableone==0.9.5
tabulate==0.9.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
tinycss2==1.4.0
torch==2.3.1
torchaudio==2.3.1
torchmetrics==1.5.2
torchvision==0.18.1
tornado==6.5.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tzlocal_1756227235/work
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1737133630106/work
wandb==0.22.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
wrapt==1.17.2
xgboost==3.0.2
xlsxwriter==3.2.9
